🧪 test(goldens): HomeScreen-Golden modernisiert (tester.view APIs); DayScreen-Golden & Firestore-Fakes entfernt (Sealed-Dart-Klassen)
